fix issue in LoadFromGRD with reading 4 bytes into a short
git-svn-id: http://vtp.googlecode.com/svn/trunk@7222 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtdata/ElevationGridIO.cpp
@@ -1180,7 +1180,7 @@ bool vtElevationGrid::LoadFromGRD(const char *szFileName,
 	  return false;
 
 	// read GRD header data
-	short nx, ny;
+	int nx, ny;
 	double xlo, xhi, ylo, yhi, zlo, zhi;
 
 	// Parse the file, determine what kind of GRD it is
